add refund

+++ src/pages/api/payment/webhook.ts
@@ -58,24 +58,15 @@ export const POST: APIRoute = async ({ request, url }) => {
                 const user = await User.getUserById(paymentResult.metadata.account_id);
                 if (user instanceof User) {
                     console.log("User Found", user);
-                    const expDate = new Date();
-                    const month = parseInt(paymentResult.additional_info?.items?.[0]?.id ?? "0");
-                    expDate.setMonth(expDate.getMonth() + month);
-                    await user.modifyMenuLimit(0).catch((err) => {
+                    await user.modifyMenuLimit(-1).catch((err) => {
                         console.error(err);
                     });
-                    const menu = new Menu({
-                        name: paymentResult.metadata.menu,
-                        active: true,
-                        description: paymentResult.metadata.description,
-                        userEmail: user.email,
-                        expDate: expDate,
-                        maxProducts: 100,
-                    })
-
-                    await menu.delete().catch((err) => {
-                        console.error(err);
-                    });
+                    const menu = await Menu.getMenuByName(paymentResult.metadata.menu);
+                    if (menu instanceof Menu) {
+                        await menu.delete().catch((err) => {
+                            console.error(err);
+                        });
+                    }
                 }
             }
         }
